backend/requirements.txt
Django>=5.0,<7.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
Pillow>=10.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
django-filter>=23.0
markdown
django-environ
drf-spectacular>=0.26.0
django-mptt>=0.15.0
django-phonenumber-field>=7.2.0
phonenumbers>=8.13.0
django-redis>=5.4.0
whitenoise>=6.6.0
django-storages>=1.14.0
boto3>=1.34.0